split up bokeh.embed.util into more meaningful pieces
