+++ bokeh/embed/standalone.py
@@ -34,10 +34,9 @@ from ..model import Model
 from ..util.compiler import bundle_all_models
 from ..util.string import encode_utf8
 from .bundle import bundle_for_objs_and_resources
-from .util import FromCurdoc
-from .util import (OutputDocumentFor, html_page_for_render_items,
-                   script_for_render_items, standalone_docs_json_and_render_items,
-                   wrap_in_onload, wrap_in_script_tag)
+from .elements import html_page_for_render_items, script_for_render_items
+from .util import FromCurdoc, OutputDocumentFor, standalone_docs_json_and_render_items
+from .wrappers import wrap_in_onload, wrap_in_script_tag
 
 #-----------------------------------------------------------------------------
 # Globals and constants
