+++ bokeh/embed/tests/test_elements.py
@@ -1,0 +1,56 @@
+#-----------------------------------------------------------------------------
+# Copyright (c) 2012 - 2017, Anaconda, Inc. All rights reserved.
+#
+# Powered by the Bokeh Development Team.
+#
+# The full license is in the file LICENSE.txt, distributed with this software.
+#-----------------------------------------------------------------------------
+
+#-----------------------------------------------------------------------------
+# Boilerplate
+#-----------------------------------------------------------------------------
+from __future__ import absolute_import, division, print_function, unicode_literals
+
+import pytest ; pytest
+
+#-----------------------------------------------------------------------------
+# Imports
+#-----------------------------------------------------------------------------
+
+# Standard library imports
+
+# External imports
+
+# Bokeh imports
+from bokeh.embed.util import RenderItem
+
+# Module under test
+import bokeh.embed.elements as bee
+
+#-----------------------------------------------------------------------------
+# Setup
+#-----------------------------------------------------------------------------
+
+#-----------------------------------------------------------------------------
+# General API
+#-----------------------------------------------------------------------------
+
+#-----------------------------------------------------------------------------
+# Dev API
+#-----------------------------------------------------------------------------
+
+class Test_div_for_render_item(object):
+
+    def test_render(self):
+        render_item = RenderItem(docid="doc123", elementid="foo123")
+        assert bee.div_for_render_item(render_item).strip() == """<div class="bk-root" id="foo123"></div>"""
+
+class Test_html_page_for_render_items(object):
+    pass
+
+class Test_script_for_render_items(object):
+    pass
+
+#-----------------------------------------------------------------------------
+# Private API
+#-----------------------------------------------------------------------------
