+++ bokeh/embed/util.py
@@ -26,17 +26,12 @@ from collections import Sequence, OrderedDict
 from contextlib import contextmanager
 
 # External imports
-from six import string_types
 
 # Bokeh imports
-from ..core.json_encoder import serialize_json
-from ..core.templates import _env, DOC_JS, FILE, MACROS, PLOT_DIV, SCRIPT_TAG
-from ..document.document import DEFAULT_TITLE, Document
+from ..document.document import Document
 from ..model import Model, collect_models
 from ..settings import settings
-from ..util.compiler import bundle_all_models
 from ..util.serialization import make_id
-from ..util.string import encode_utf8, escape, indent
 
 #-----------------------------------------------------------------------------
 # Globals and constants
@@ -197,85 +192,6 @@ def submodel_has_python_callbacks(models):
 
     return has_python_callback
 
-def div_for_render_item(item):
-    '''
-        item: RenderItem
-    '''
-    return PLOT_DIV.render(doc=item, macros=MACROS)
-
-def html_page_for_render_items(bundle, docs_json, render_items, title,
-                               template=None, template_variables={}):
-    '''
-
-    '''
-    if title is None:
-        title = DEFAULT_TITLE
-
-    bokeh_js, bokeh_css = bundle
-
-    json_id = make_id()
-    json = escape(serialize_json(docs_json), quote=False)
-    json = wrap_in_script_tag(json, "application/json", json_id)
-
-    script = bundle_all_models()
-    script += script_for_render_items(json_id, render_items)
-    script = wrap_in_script_tag(script)
-
-    context = template_variables.copy()
-
-    context.update(dict(
-        title = title,
-        bokeh_js = bokeh_js,
-        bokeh_css = bokeh_css,
-        plot_script = json + script,
-        docs = render_items,
-        base = FILE,
-        macros = MACROS,
-    ))
-
-    if len(render_items) == 1:
-        context["doc"] = context["docs"][0]
-        context["roots"] = context["doc"].roots
-
-    # XXX: backwards compatibility, remove for 1.0
-    context["plot_div"] = "\n".join(div_for_render_item(item) for item in render_items)
-
-    if template is None:
-        template = FILE
-    elif isinstance(template, string_types):
-        template = _env.from_string("{% extends base %}\n" + template)
-
-    html = template.render(context)
-    return encode_utf8(html)
-
-def script_for_render_items(docs_json_or_id, render_items, app_path=None, absolute_url=None):
-    '''
-
-    '''
-    if isinstance(docs_json_or_id, string_types):
-        docs_json = "document.getElementById('%s').textContent" % docs_json_or_id
-    else:
-        # XXX: encodes &, <, > and ', but not ". This is because " is used a lot in JSON,
-        # and encoding it would significantly increase size of generated files. Doing so
-        # is safe, because " in strings was already encoded by JSON, and the semi-encoded
-        # JSON string is included in JavaScript in single quotes.
-        docs_json = serialize_json(docs_json_or_id, pretty=False) # JSON string
-        docs_json = escape(docs_json, quote=("'",))               # make HTML-safe
-        docs_json = docs_json.replace("\\", "\\\\")               # double encode escapes
-        docs_json =  "'" + docs_json + "'"                        # JS string
-
-    js = DOC_JS.render(
-        docs_json=docs_json,
-        render_items=serialize_json([ item.to_json() for item in render_items ], pretty=False),
-        app_path=app_path,
-        absolute_url=absolute_url,
-    )
-
-    if not settings.dev:
-        js = wrap_in_safely(js)
-
-    return wrap_in_onload(js)
-
 class RenderRoot(object):
 
     def __init__(self, elementid, id, name=None, tags=None):
@@ -419,43 +335,9 @@ def standalone_docs_json_and_render_items(models):
 
     return (docs_json, render_items)
 
-def wrap_in_onload(code):
-    '''
-
-    '''
-    return _ONLOAD % dict(code=indent(code, 4))
-
-def wrap_in_safely(code):
-    '''
-
-    '''
-    return _SAFELY % dict(code=indent(code, 2))
-
-def wrap_in_script_tag(js, type="text/javascript", id=None):
-    '''
-
-    '''
-    return SCRIPT_TAG.render(js_code=indent(js, 2), type=type, id=id)
-
 #-----------------------------------------------------------------------------
 # Private API
 #-----------------------------------------------------------------------------
-
-_ONLOAD = """\
-(function() {
-  var fn = function() {
-%(code)s
-  };
-  if (document.readyState != "loading") fn();
-  else document.addEventListener("DOMContentLoaded", fn);
-})();\
-"""
-
-_SAFELY = """\
-Bokeh.safely(function() {
-%(code)s
-});\
-"""
 
 #-----------------------------------------------------------------------------
 # Code
