+++ bokeh/embed/wrappers.py
@@ -1,0 +1,84 @@
+#-----------------------------------------------------------------------------
+# Copyright (c) 2012 - 2017, Anaconda, Inc. All rights reserved.
+#
+# Powered by the Bokeh Development Team.
+#
+# The full license is in the file LICENSE.txt, distributed with this software.
+#-----------------------------------------------------------------------------
+'''
+
+'''
+
+#-----------------------------------------------------------------------------
+# Boilerplate
+#-----------------------------------------------------------------------------
+from __future__ import absolute_import, division, print_function, unicode_literals
+
+import logging
+log = logging.getLogger(__name__)
+
+#-----------------------------------------------------------------------------
+# Imports
+#-----------------------------------------------------------------------------
+
+# Standard library imports
+
+# External imports
+
+# Bokeh imports
+from ..core.templates import SCRIPT_TAG
+from ..util.string import indent
+
+#-----------------------------------------------------------------------------
+# Globals and constants
+#-----------------------------------------------------------------------------
+
+#-----------------------------------------------------------------------------
+# General API
+#-----------------------------------------------------------------------------
+
+#-----------------------------------------------------------------------------
+# Dev API
+#-----------------------------------------------------------------------------
+
+def wrap_in_onload(code):
+    '''
+
+    '''
+    return _ONLOAD % dict(code=indent(code, 4))
+
+def wrap_in_safely(code):
+    '''
+
+    '''
+    return _SAFELY % dict(code=indent(code, 2))
+
+def wrap_in_script_tag(js, type="text/javascript", id=None):
+    '''
+
+    '''
+    return SCRIPT_TAG.render(js_code=indent(js, 2), type=type, id=id)
+
+#-----------------------------------------------------------------------------
+# Private API
+#-----------------------------------------------------------------------------
+
+_ONLOAD = """\
+(function() {
+  var fn = function() {
+%(code)s
+  };
+  if (document.readyState != "loading") fn();
+  else document.addEventListener("DOMContentLoaded", fn);
+})();\
+"""
+
+_SAFELY = """\
+Bokeh.safely(function() {
+%(code)s
+});\
+"""
+
+#-----------------------------------------------------------------------------
+# Code
+#-----------------------------------------------------------------------------
